Add a workaround for a VS 2015 bug
It appears that VS 2015 is unable to template some wrapped pointer types
correctly. In particular, while it could normally handle `T*`, it is
unable to handle say `std::shared_ptr<T>`. This patch adds a workaround
in VIGRA by overloading `get_pointer` for different types of `T`. By
doing this, VS 2015 is able to build through these issues. The patch has
already been accepted in VIGRA master and should be available in the
next release of VIGRA.

ref: https://msdn.microsoft.com/vs-knownissues/vs2015-update3#other-issues