Patch Windows NMake Python test run
The vigranumpy test run appears to be adding a directory to check for
`nose` tests, which is not present during the NMake run. Though this
directory is present for Visual Studio test runs. Hence this patch
adjust which directory and how it is added to ensure that the Visual
Studio and NMake cases both work.