Move `cmake` into `build`
As `cmake` is a build tool and we are not linking against (nor should we
be constrained by its runtime dependency constraints), move it out of
the `host` section into the `build` section. That way we can still call
it on the command line, but it does not contaminate our build
environment.