Specify Boost Python library to use
Start specifying the Boost Python library to use as it appears this
changes every Boost version and it is too much work to patch VIGRA for
each change. If we keep the change in the recipe at least it will be
easier to update and reflect what any specific Boost Python version
needs at least.