Backport exception declaration fix for C++11
In C++11, the old `throw` exception declaration was deprecated and
replaced with `noexcept(false)` in N3051. As of C++17, the old `throw`
style exception declaration was dropped in P0003R5. As Linux builds are
compiled against C++17, this is blocking a build with the new compilers.
To fix this we backport this upstream patch, which handles this change
by selectively using the right behavior between the different C++
versions; using the non-deprecated behavior as of C++11 where it first
appears.

ref: http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2010/n3051.html
ref: http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2016/p0003r5.html